fix(clean): code cleanup
