--- components/CenteredPanel.module.scss
@@ -1,6 +1,0 @@
-.panel {
-    width: 100%;
-    height: 100vh;
-    display: grid;
-    align-items: center;
-}
