--- components/FixedBanner.module.scss
@@ -1,17 +1,0 @@
-.banner {
-    position: fixed;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    height: 100vh;
-    max-width: 100%;
-    @media screen and (max-width: 760px) {
-        max-width: 10vw;
-    }
-}
-
-.image {
-    @media screen and (max-width: 760px) {
-        max-width: 10vw;
-    }
-}
