+++ components/HexPanel.js
@@ -3,14 +3,28 @@ import styles from "./HexPanel.module.scss";
 
 export default function HexPanel(props) {
   return (
-    <Box
-     className={styles.panel}
-    >
+    <Box className={styles.panel}>
       {props.children}
-      <Image src="/hex.svg" sx={{ position: `absolute`, left: `5px`, top: `5px` }} />
-      <Image src="/hex.svg" sx={{ position: `absolute`, left: `5px`, bottom: `5px` }} />
-      <Image src="/hex.svg" sx={{ position: `absolute`, right: `5px`, top: `5px`  }} />
-      <Image src="/hex.svg" sx={{ position: `absolute`, right: `5px`, bottom: `5px`  }} />
+      <Image
+        src="/assets/hex.svg"
+        sx={{ position: `absolute`, left: `5px`, top: `5px` }}
+        alt={"hex?"}
+      />
+      <Image
+        src="/assets/hex.svg"
+        sx={{ position: `absolute`, left: `5px`, bottom: `5px` }}
+        alt={"hex?"}
+      />
+      <Image
+        src="/assets/hex.svg"
+        sx={{ position: `absolute`, right: `5px`, top: `5px` }}
+        alt={"hex?"}
+      />
+      <Image
+        src="/assets/hex.svg"
+        sx={{ position: `absolute`, right: `5px`, bottom: `5px` }}
+        alt={"hex?"}
+      />
     </Box>
   );
 }
