+++ components/drink/FixedBanner.js
@@ -1,13 +1,13 @@
 import { Box, Image } from "@chakra-ui/react";
 import styles from "./FixedBanner.module.scss";
 
-export default function FixedBanner(props, logo) {
+export default function FixedBanner(props) {
   return (
     <Box className={styles.banner} sx={{ zIndex: `4` }}>
       <Image
-        src={logo}
+        src={props.logo}
         className={styles.image}
-        alt={logo}
+        alt={"logo"}
         sx={{ maxHeight: `90vh` }}
       />
     </Box>
