--- components/drink/HexPanel.module.scss
@@ -1,17 +1,0 @@
-.panel {
-    position: relative;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    border: 5px solid #ebebeb;
-    border-radius: 15px;
-    padding: 35px;
-    width: 80vw;
-    place-items: center;
-    height: 50vh;
-    @media screen and (max-width: 760px) {
-        grid-template-columns: 1fr;
-        width: 100%;
-        height: fit-content;
-        row-gap: 1rem;
-    }
-}
