+++ pages/api/copy.json
@@ -6,7 +6,11 @@
     "beer": "From below his sacrificial altar, we drove our blade into the demon’s flesh. The harvest of battle is blood, intense, complex and rich, with flavors of roasted malts, dark chocolate, dark fruits and espresso. Drink its poison, all 14% abv of it!",
     "callToAction": "Summon Moloch",
     "traits": { "abv": 14, "ibu": 24 },
-    "lore": "Through the long winter of '18 and the mania that followed, the mercenaries at Raid Guild have gathered to share stories of battle, cleaning the Blood of Moloch from their armor and sharpening their swords. Inspired by the tirelessness of fellow Raid Guild members, the noble Sasquatch, a cleric and master of potions, concocted a fermented beverage to quench the thirsts for our ongoing battle with Moloch. The recipe and the brewing of this dark nectar has resulted in what will forever be known as Blood Of Moloch."
+    "lore": {
+      "first": "Through the long winter of '18 and the mania that followed, the mercenaries at Raid Guild have gathered to share stories of battle, cleaning the Blood of Moloch from their armor and sharpening their swords. ",
+      "second": "Inspired by the tirelessness of fellow Raid Guild members, the noble Sasquatch, a cleric and master of potions, concocted a fermented beverage to quench the thirsts for our ongoing battle with Moloch. The recipe and the brewing of this dark nectar has resulted in what will forever be known as Blood Of Moloch."
+    },
+    "canImage": "https://res.cloudinary.com/raidguild/image/upload/v1643180188/BroodGuild/misc/one_can_la3mhf.svg"
   },
   "red-pil": {
     "name": "Red Pil",
@@ -15,6 +19,10 @@
     "beer": "Crisp and dry, traditional pilsner profile with a touch of lemongrass, bergamot and blacktea from late Saphir hop additions. Fermented with OldWorld+NewBeast Kveik Yeast. All the clean pilsner taste, with only 1/6 the brewing time.",
     "traits": { "abv": 5.1, "ibu": 50 },
     "callToAction": "Take the Red Pil",
-    "lore": "When you know there’s got to be a better way and you’re longing to follow the white rabbit, a sip of this crisp summer brew is like soft lips near your ear, encouraging you to choose the Red Pil."
+    "lore": {
+      "first": "When you know there’s got to be a better way and you’re longing to follow the white rabbit, a sip of this crisp summer brew is like soft lips near your ear, encouraging you to choose the Red Pil.",
+      "second": ""
+    },
+    "canImage": "https://res.cloudinary.com/raidguild/image/upload/v1643180188/BroodGuild/misc/one_can_la3mhf.svg"
   }
 }
